Updated Codeception to ^5 and replaced `zendframework/zend-diactoros` with `nyholm/psr7`
