+++ codeception.yml
@@ -1,13 +1,11 @@
 actor: Tester
 paths:
     tests: tests
-    log: tests/_output
+    output: tests/_output
     data: tests/_data
     support: tests/_support
     envs: tests/_envs
 settings:
-    bootstrap: _bootstrap.php
-    colors: false
     memory_limit: 1024M
 extensions:
     enabled:
