+++ tests/Fixtures/Bar.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Kodus\Logging\Test\Fixtures;
+
+class Bar extends Foo
+{
+    public $bat = "BAT";
+}
