+++ tests/Fixtures/Baz.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Kodus\Logging\Test\Fixtures;
+
+class Baz
+{
+    public $foo;
+    public $bar;
+    public $baz;
+
+    public function __construct()
+    {
+        $this->foo = new Foo();
+        $this->bar = new Bar();
+        $this->baz = $this;
+    }
+}
