+++ tests/Fixtures/Foo.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Kodus\Logging\Test\Fixtures;
+
+class Foo
+{
+    public $foo = "FOO";
+    protected $bar = "BAR";
+    private $baz = "BAZ";
+}
