--- tests/_bootstrap.php
@@ -1,2 +1,0 @@
-<?php
-// This is global bootstrap for autoloading
