--- tests/unit.suite.dist.yml
@@ -1,5 +1,0 @@
-class_name: UnitTester
-modules:
-    enabled:
-        - Asserts
-        - Helper\Unit
