--- tests/unit/_bootstrap.php
@@ -1,2 +1,0 @@
-<?php
-// Here you can initialize variables that will be available to your tests
